split the app into auth and unauth parts
