+++ views/src/app/Auth/pages/Home/Home.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Nav from '../../components/navbar/Navbar'
+
+const Home = () => {
+    return (
+        <div>
+            <div>HOME PAGE FOR AUTHORIZED USER</div>
+        </div>
+    )
+}
+
+export default Home;
