+++ views/src/app/Unauth/components/login/Login.js
@@ -1,48 +1,40 @@
 import React, { useState } from "react";
 import { Button, FormGroup, FormControl, ControlLabel } from "react-bootstrap";
 import { withRouter } from 'react-router-dom';
+import 'bootstrap/dist/css/bootstrap.min.css';
 import axios from 'axios';
+import "./Login.css";
 
-const RegisterForm = props => {
-    const [username, setUsername] = useState("");
+const LoginForm = props => {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-    const { history } = props;
+    const { history, loginHandler } = props;
 
-    let creatUser = async () => {
+    let logInUser = async () => {
         try {
-            console.log(username, email, password);
-            const res = await axios.post('http://localhost:3001/api/users/', { username, email, password });
+            const res = await axios.post('http://localhost:3001/api/users/login', { email, password });
             if (res.status === 200) {
-                console.log("Successfully creating new user");
-                history.push('/profile');
+                console.log("User log in successful");
                 localStorage.token = res.data.token;
+                localStorage.username = res.data.username;
+                history.push('/profile');
+                loginHandler();
             } else {
                 console.log("FAIL log in"); //TODO
             }
         } catch (err) {
-            console.log("FAIL creating new user: ", err);
+            console.log("Failed to send request:", err);
         }
     }
 
     let handleSubmit = event => {
         event.preventDefault();
-        creatUser();
+        logInUser();
     }
 
     return (
         <div className="Login">
             <form onSubmit={handleSubmit}>
-                <FormGroup controlId="username" bsSize="large">
-                    <ControlLabel>Username</ControlLabel>
-                    <FormControl
-                        autoFocus
-                        type="username"
-                        value={username}
-                        onChange={e => setUsername(e.target.value)}
-                        required
-                    />
-                </FormGroup>
                 <FormGroup controlId="email" bsSize="large">
                     <ControlLabel>Email</ControlLabel>
                     <FormControl
@@ -63,11 +55,11 @@ const RegisterForm = props => {
                     />
                 </FormGroup>
                 <Button block bsSize="large" type="submit">
-                    Create Account
+                    Login
                 </Button>
             </form>
         </div>
     );
 }
 
-export default withRouter(RegisterForm);
+export default withRouter(LoginForm);
