+++ views/src/app/Unauth/components/navbar/Navbar.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import { Link, withRouter } from 'react-router-dom';
-import './Navbar.css';
 
-const Nav = ({ location }) => {
+const Nav = ({ location, logoutHandler }) => {
     return (
         <div className="nav_menu">
             <img />
+            Meh meh meh
             <Link className={`nav_item ${location.pathname === '/' ? 'active' : ""}`} to='/'>Home
             </Link>
-            <Link className={`nav_item ${location.pathname === '/login' ? 'active' : ""}`} to='/login'>Log In
-            </Link>
-            <Link className={`nav_item ${location.pathname === '/register' ? 'active' : ""}`} to='/register'>Register
+            <Link className={`nav_item ${location.pathname === '/posts' ? 'active' : ""}`} to='/posts'>Posts
             </Link>
         </div>
     )
