+++ views/src/app/Unauth/components/register/Register.js
@@ -1,39 +1,49 @@
 import React, { useState } from "react";
 import { Button, FormGroup, FormControl, ControlLabel } from "react-bootstrap";
 import { withRouter } from 'react-router-dom';
-import 'bootstrap/dist/css/bootstrap.min.css';
 import axios from 'axios';
-import "./Login.css";
 
-const LoginForm = props => {
+const RegisterForm = props => {
+    const [username, setUsername] = useState("");
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-    const { history } = props;
+    const { history, loginHandler } = props;
 
-    let logInUser = async () => {
+    let creatUser = async () => {
         try {
-            const res = await axios.post('http://localhost:3001/api/users/login', { email, password });
+            console.log(username, email, password);
+            const res = await axios.post('http://localhost:3001/api/users/', { username, email, password });
             if (res.status === 200) {
-                console.log("User log in successful");
-                localStorage.token = res.data.token;
-                localStorage.username = res.data.username;
+                console.log("Successfully creating new user");
                 history.push('/profile');
+                localStorage.token = res.data.token;
+                loginHandler();
             } else {
                 console.log("FAIL log in"); //TODO
             }
         } catch (err) {
-            console.log("Failed to send request:", err);
+            console.log("FAIL creating new user: ", err);
         }
     }
 
     let handleSubmit = event => {
         event.preventDefault();
-        logInUser();
+        creatUser();
     }
 
     return (
         <div className="Login">
             <form onSubmit={handleSubmit}>
+                <FormGroup controlId="username" bsSize="large">
+                    <ControlLabel>Username</ControlLabel>
+                    <FormControl
+                        autoFocus
+                        type="username"
+                        value={username}
+                        onChange={e => setUsername(e.target.value)}
+                        required
+                    />
+                </FormGroup>
                 <FormGroup controlId="email" bsSize="large">
                     <ControlLabel>Email</ControlLabel>
                     <FormControl
@@ -54,11 +64,11 @@ const LoginForm = props => {
                     />
                 </FormGroup>
                 <Button block bsSize="large" type="submit">
-                    Login
+                    Create Account
                 </Button>
             </form>
         </div>
     );
 }
 
-export default withRouter(LoginForm);
+export default withRouter(RegisterForm);
