--- views/src/pages/home/Home.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import LoginForm from '../../components/login/Login';
-import RegisterForm from '../../components/reg/Register';
-
-const Home = () => {
-    return (
-        <div>
-            <LoginForm />
-            <RegisterForm />
-        </div>
-    )
-}
-
-export default Home;
